Additional UWP preparations
