+++ Windows/DSoundStream.cpp
@@ -4,9 +4,11 @@
 #include <process.h>
 
 #include "thread/threadutil.h"
+#include "Common/OSVersion.h"
 #include "Core/Reporting.h"
 #include "Core/Util/AudioFormat.h"
 #include "Windows/W32Util/Misc.h"
+#include "Common/OSVersion.h"
 
 #include "dsoundstream.h"
